+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Array equals method
 - AABB containment and intersection methods
 - `Array.atParam()` method
+- Path operations needed in PathLite
 
 ### Changed
 - RNG methods are spelled out completely.